Move contents of detail header to separate component
